When dropping privileges, enable and use the xrdhttp-pelican plugin
The plugin provides a means for an unprivileged process (namely, `pelican`)
to request xrootd send itself a signal via a Unix socket and to copy
the host cert via a sent file descriptor.

This refactors the launcher kill logic so the plugin can be used instead
of the traditional signal-based approach.